[WIP] Adding a TimeTravel improver for nginx
---------------------NOT READY FOR REVIEW YET--------------------

This is not the best way. Maybe we want a generic time travel improver
which could run on multiple data source. This is highly experimental and
is meant to serve as proof of work for current model.
Currently, this is incomplete and doesn't work. The classes and
functions used in time travel have not been ported properly.

This commit might need a full rewrite

Signed-off-by: Hritik Vijay <hritikxx8@gmail.com>